Moved SwipeHandler to formentry package
